# Expand locking in GetResult methods