`GetResult` calls `ObserveCancellation` that modifies state. Decreases performance but we need it.

Before:
```
                  Method |        Mean |    StdErr |     StdDev |          RPS |
------------------------ |------------ |---------- |----------- |------------- |
 ParseLiveAspNetTwoTasks | 291.2868 ns | 2.0583 ns | 11.2738 ns | 3,433,043.01 |
   ParseLiveAspNetInline | 204.2358 ns | 2.8546 ns | 15.6355 ns | 4,896,300.75 |

```

After:

```
                  Method |        Mean |    StdErr |     StdDev |          RPS |
------------------------ |------------ |---------- |----------- |------------- |
 ParseLiveAspNetTwoTasks | 326.4556 ns | 3.3105 ns | 18.1321 ns | 3,063,203.49 |
   ParseLiveAspNetInline | 237.0526 ns | 3.8329 ns | 20.9937 ns | 4,218,472.53 |
```

@davidfowl